+++ src/plugins/reference_analysis/jsenv_plugin_reference_analysis.js
@@ -104,7 +104,9 @@ const jsenvPluginReferenceAnalysisInclude = ({
       const protocolIsSupported = supportedProtocols.some(
         (supportedProtocol) => protocol === supportedProtocol,
       );
-      reference.mustIgnore = !protocolIsSupported;
+      if (!protocolIsSupported) {
+        reference.mustIgnore = true;
+      }
     },
   };
 };